requirements.txt
python-dotenv
pydantic
google-generativeai
google-api-python-client
mem0
streamlit